MDEV-515 innodb bulk insert
- Introduced assign_stat_n_rows() in dict_table_t. It calculates the
number of rows in the table and assign stat_n_rows during
ha_innobase::open()

- Introduced empty_table() in dict_table_t. Basically it empties all the
indexes associated with table (not covered the fts index). This is
undo operation of bulk operation.

- Introduced new variable bulk_trx_id in dict_table_t. It stores the
transaction id of bulk insert. Basically it is protected by exclusive
lock of the table.

- Introduced new variable non_empty in dict_table_t. It indicates the
whether the table is empty.

- If table is empty then INSERT, INSERT..SELECT does take exclusive lock
on the table.

- Introduced new undo log record "TRX_UNDO_UNEMPTY". It should be first
undo log during bulk operation. While rollback, if innodb encounters
the undo record then it should empty the table.

Limitations:
===========
- Parallel read should give empty table depends on bulk_trx_id
- Need to handle the online alter when the table is empty.